refactor(mpp): move exec_mpp_worker into aggregatescan/mpp.rs submodule
aggregatescan/mod.rs carried ~360 LOC of MPP worker dispatch in
exec_mpp_worker (logical-plan deserialization, distributed-physical-plan
build, fragment discovery, join_all async machinery). The rest of mod.rs
is the leader-side trait impl + serial Tantivy paths; the MPP worker
exec is functionally a separate concern.

Move exec_mpp_worker into aggregatescan/mpp.rs as a pub(super) method
on `impl AggregateScan`. mod.rs picks it up via `pub mod mpp;`. The
cross-impl-block call from `exec_custom_scan` continues to work.
build_mpp_session_context and stash_mpp_plan_bytes stay in mod.rs for
now — they're also called from non-MPP-worker paths.

Net: mod.rs loses 352 LOC; new mpp.rs is 386 LOC. The unused-imports
sweep the move exposed in mod.rs (RuntimeEnvBuilder, list_segment_ids,
etc.) now lives where the code that uses them lives.
